ai_services/set_training_translation/requirements.txt
# 手语翻译系统依赖包
# 安装命令: pip install -r requirements.txt

# 深度学习框架
tensorflow==2.15.0

# 计算机视觉和手部跟踪
mediapipe==0.10.9
opencv-python==4.8.1.78

# 数值计算
numpy==1.26.4

# Web 服务器
flask==3.0.0
flask-cors==4.0.0

# 图像处理
pillow==10.1.0

# 机器学习工具
scikit-learn==1.3.2

# TensorFlow 依赖
protobuf==3.20.3
jax==0.4.13
jaxlib==0.4.13

